requirements.txt
google-api-python-client==2.129.0
google-auth-oauthlib==1.2.0
notion-client==2.2.1
pip-chill==1.0.3
python-dotenv==1.0.1
